Finished Product
This is the final revision of my current ISP project. It a game of Pong
programmed in Swift using the Apple Xcode Application. It currently has
the ‘paddles’ running automatically, but if the two lines running that
function are commented out, it will run based on player input from an
Arduino device via Serial Communication.